+++ packages/did-client/src/did-client.ts
@@ -4,7 +4,7 @@ import { default as VeridaWallet } from "./wallet"
 import { getResolver } from '@verida/vda-did-resolver'
 import { VdaDid } from '@verida/vda-did'
 import { Resolver } from 'did-resolver'
-import { Web3CallType, DIDClientConfig, VdaDidEndpointResponses, Web3ResolverConfigurationOptions, Web3SelfTransactionConfig, Web3MetaTransactionConfig, VeridaWeb3ConfigurationOptions, Web3SelfTransactionConfigPart, IDIDClient } from "@verida/types"
+import { Web3CallType, DIDClientConfig, VdaDidEndpointResponses, Web3ResolverConfigurationOptions, Web3SelfTransactionConfig, Web3MetaTransactionConfig, VeridaWeb3ConfigurationOptions, Web3SelfTransactionConfigPart, IDIDClient, VeridaDocInterface } from "@verida/types"
 
 export class DIDClient implements IDIDClient {
 
@@ -194,6 +194,6 @@ export class DIDClient implements IDIDClient {
             throw new Error(`DID resolution error (${resolutionResult.didResolutionMetadata.error}): ${resolutionResult.didResolutionMetadata.message} (${did})`)
         }
 
-        return <DIDDocument> resolutionResult.didDocument
+        return new DIDDocument(<VeridaDocInterface> resolutionResult.didDocument)
     }
 }